Use a different data structure for storing graph state
Workflow graph can have cycles. In such case, a task state will be overwritten. This patch separates the execution graph from the execution state. There's a new task flow dictionary that keeps track of task execution sequence and which state is current for tasks in cycle. Additional methods are included to work with the task flow dictionary.