Implement state machine for task state transition
Implement a simple state machine that process action execution event for task state transition. The workflow conductor serializes and deserializes often and so the state machine needs to be light weight that does not require constant reloading and can accomodate the frequent serialization. A class method is used to process the action execution event. The method looks up the mapping of the task state transition from a static dictionary.