Implement state machine for workflow state transition
Implement a simple state machine that process task execution event for workflow state transition. The workflow conductor serializes and deserializes often and so the state machine needs to be light weight that does not require constant reloading and can accomodate the frequent serialization. A class method is used to process the task execution event. The method looks up the mapping of the workflow state transition from a static dictionary. During processing, additional context information about the state of the workflow is appended to the event. The context include information such as whether there are other tasks that are active, if any tasks are paused or canceled, and if the workflow still has tasks to execute.