Add with items support to native task spec
Add a property named `with` to the native task spec in support of with items. The `with` property in the json schema is type of string with a regular expression to match the patterns "x in <expression>" or "x, y, ... in <expression>". Add unit tests to cover validation of various patterns.