Refactor finalize_context in task spec to get the transition precisely
Refactor finalize_context in task spec to require addition task transition information as input. For the native model, the task spec uses the ref attribute stored in the edge's list of attributes to identify the task transition spec precisely. Previously, the method compares the criteria string which is not precise and bug prone.