Process workflow output in other completed states
Previously, workflow output is processed only when workflow succeeds. This patch enables workflow output to be processed when workflow completes, regardless of success or failure. If there is a evaluation error, the line item in the output will be skipped and evaluation error recorded. When there is an output evaluation error, the workflow will be marked as failed.