Fix conductor get next tasks for with items task with empty list
Fix a condition for with items task where the items list is empty. Before the patch, when the conductor get next tasks, the with items task will not be included in next tasks because the list is empty. The conductor is refactored to return the with items task in next tasks but without any action specs. The service using conductor will have to handle this accordingly to complete the with items task appropriately. The reasoning here is that the service such as StackStorm also maintains DB records for the task execution. Therefore, the conductor cannot auto forward to the next task but let the service manually handle the task execution even though the items list is empty and then manually forward the workflow progress.