Value in quotes in shorthand publish should be evaluated as string type Fixes issue https://github.com/StackStorm/orquesta/issues/130
During `parse_inline_params` is called to parse parameters input, the leading and trailing single/double quotes is removed. Digit and True/False in quotes will be parsed as number and boolean type instead of String after `json.loads()`.

Solution:
For string in quotes (not includes dictionary  `'x=\'{\"a\": 1}\''`), 'json.loads' is not called.