Refactor workflow execution graph
Refactor the native composer to use the original graph from the workflow definition for the execution graph. Previously, the composer expands non-join tasks that are referenced more than once and duplicate subgraphs in the execution graph (aka split). The result of the expansion led to an inflated graph with exponentially growing number of task nodes and task transitions. The refactored composer keeps track of route details and route index when conducting the workflow execution. When a split is detected, the route change is recorded in the conductor internal state and a new route index is assigned. As part of the refactoring, the structure of the internal workflow execution state data for a task execution is enhanced to include references to previous and next set of tasks.