Refactor the base class for the unit test modules
Refactor the base class for the unit test modules to simplify operations to change task states and to assert next task(s). Refactor the method that assert conducting sequences to include checking route index for task sequences, checking on details of routes, and checking the final workflow state in which it wasn't checking.