Refactor task_state function to resolve task correctly
For the split use case, the routing information before and after a split is different. The task_state function need to be able to correctly identify the correct task entry just by task name. The task_state function is refactored to look use the routing information for the current task and then recursively look up the task in earlier associated route segments.