Fix overwritten context variables on task join
Refactor the internal data structure for storing task context. Previously, if there is change in context at task transition after publish, an entire copy of the context is stored. So when a task join happens and one branch modifies one variable and another branch modifies another variable, the last branch that merges the context for the join task will overwrite the variables for the other branch. With this change, only diff in the context is stored after each task transition and the join task will merge the diffs chronologically. This will also reduce the memory footprint used by the internal data structure because only diffs are stored.