Rename state to status
Rename action, task, and workflow execution state to status. The renaming is part of code cleanup so it is consistent with the use of execution status in StackStorm. Also, this change is in preparation for the refactoring of the conductor internal state and want to avoid confusion with the use of the term.