Refactor concept of task flow to workflow and task state
The use of TaskFlow and associated terms are not immediately clear. Therefore, refactor TaskFlow to WorkflowState. Move the _workflow_status variable in the WorkflowConductor into WorkflowState. Refers to task entry in the list of task execution sequences as task_state_entry where it was task_flow_entry previously. Refactor the WorkflowStateMachine and TaskStateMachine and pass the WorkflowState as input instead of the WorkflowConductor. Conceptually, the WorkflowConductor uses the WorkflowState to conduct the workflow execution and passes the WorkflowState into the state machines for workflow and task status transition.