Fix task transition error on completing task in pending status
A workflow goes into paused status when a task goes into pending status. If the task transition has error, the workflow does not go into failed status when the task is completed. This patch add an entry in the state machine for workflow to go from paused to failed status.