Fix Jinja evaluator to evaluate block expression correctly
If there is a for loop in the Jinja block expression and the item is a complex data type, the current Jinja evaluator converted the item into string type. So if the item is a dictionary, and user tries to access some key value pair, the Jinja evaluator returns an error saying the item is undefined. The Jinja evaluator tries to evaluate individual expression first, even for expression within the for loop, before evaluating the block expression which resulted in the undefined error when the item of the for loop is referenced. To fix, the text is checked to see if there is a Jinja block expression. If there is a Jinja block expression, the entire text is rendered together. Otherwise, if there is no Jinja block expression, then standalone expressions will be rendered individually to preserve type. The method to render Jinja block expression returns string type only.