Fix conducting of cycle with a fork
If there is a branch that fork from a cycle in the workflow graph, during runtime and the workflow runs the cycle more than once, tasks under the forked branch will not execute a second time. The issue is caused by the networkx graph not recognizing the forked branch as part of the cycle. The conductor is changed to determine a task is executed more than once as part of the cycle when the last task entry is completed and the new task status is one of the starting execution statuses.