Add new engine command named "continue"
Add a new engine command named "continue". Previously, the default value when `do` is not specified is `noop`. When the previous task state is one of the failure states, noop will be assumed as remediation and the workflow execution will succeed. With `continue`, it will be replaced as the default value when `do` is not specified. The workflow engine will not alter the previous task state and will continue to conduct the workflow execution. If the previous task state is one of the failure states, the conductor will continue and fail the workflow execution.