Refactor conductor to not store each item result in task state
Refactor conductor to not store individual item result for a with items task in the task state. If there are a lot of items and/or result size is huge per item, then this will be a performance impact on database write operations when recording the conductor state. A new TaskItemActionExecutionEvent is introduced to distinguish between a normal action execution event and one for an item. The accumulated result is only stored temporary in the conductor state and evaluated by the conductor when all the items have completed execution.