Separate render workflow output from update task state
In certain use case, the service using orquesta gets the final task status only after updating the task state. If the rendering of workflow output on completion of the workflow execution calls a task related function to get task status and result such as the `task` function in StackStorm, the render would return an inaccurate task state. To address this issue, the rendering of workflow output is not done by default when updating the task state. The services consuming orquesta need to render workflow output separately and explicitly. The service should update task state, then check the workflow status before rendering the workflow output.