Handle task retry in workflow conductor
Refactor workflow conductor to handle task retry. The task retry is implemented as an engine event. The state machines handles the normal action execution event for the task as normal to determine new task status. After the new task status is determined and there is a retry defined for the task, then the workflow conductor evaluate whether the task need to be retried. This retry evaluation occurs after the task status is determined because the retry when expression may required the task status and task result for evaluation. The retry event is implemented as an engine event because the retry is decided by the conductor and it is just notifying the state machines of this retry event. The alternative is to push the decision making to the state machines, which was the original implementation. The decision requires quite a bit of context and requires pushing more data to the state machines which is not preferred.