Support of retry as an engine command in task transition
Add support to allow the keyword "retry" as an engine command in task transitions (i.e. do: retry). The default count for the retry is set to 3 with no delay and if the when condition is not given, it will be defaulted to "<% completed() %>" to be consistent with normal task transition.