Fix issue with join on complete of previous tasks
Given one or more tasks that joins on complete but then some of the tasks failed, the workflow state machine determines the workflow failed because the inbound criteria for the join task is not satisfied and there are no more tasks to run. An option is added to skip evaluation of the inbound criteria for the join task during evaluation in the workflow state machine. The workflow state machine will identify that there is a potential next task to run and keep the workflow state as running.