Fix workflow state when one or more items failed in a with items task
When one or more items failed in a with items task, the with items task is put back into staging to support manual rerun. However, this impacted the determination of workflow status because the workflow thinks there are still tasks in staging that need to run. The patch update the has_staged_tasks function in WorkflowState to ignore with items task that were put back in staging.