Fix join where inbound transition has failed condition
For join, if an inbound task fails as defined in the task transition, join sometimes is not satisfied. This issue occurs if there are multiple inbound tasks and the inbound task that failed occurs before other inbound tasks complete. The root cause is the next tasks evaluation in the state machine. Since the inbound task fails and other inbound tasks are still running, the join task is not ready yet and the state machine thought there are no next tasks and fail the workflow. The fix here adds a check to see if the next task is a join task and act appropriately. Also, the rehearsal engine is also refactored to launch all tasks in the queue into running state to simulate concurrency. Previously, it was launching and complete tasks serially and missed test coverage for this use case.